Tighten validation of `stop_after` argument to algorithms (#1071)
This makes the interpretation of `stop_after` when passed to solvers more strict:
* `stop_after = 0` or negative values are not accepted.
* `stop_after = 0` was previously used to mean "no limit".  This was an accidental artefact of stopAfter = 0 meaning this in C++.  Replaced with `std::optional` in C++ to make the intent of this parameter more clear.
* Do not accept `bool` for this parameter in Python as that could lead to surprising behaviour.

---------

+++ src/solvers/enumpoly/polysolver.cc
@@ -181,7 +181,7 @@ Vector<double> PolynomialSystemSolver::ImprovingNewtonStep(const Vector<double> 
 }
 
 std::list<Vector<double>> PolynomialSystemSolver::FindRoots(const Rectangle<double> &r,
-                                                            const int max_roots,
+                                                            const size_t max_roots,
                                                             const size_t max_rectangles,
                                                             bool &p_budgetExceeded,
                                                             const CancelToken &p_cancel)
